--- openwillis/features/api.py
@@ -1,16 +1,0 @@
-# author:    Vijay Yadav
-# website:   http://www.bklynhlth.com
-
-# import the required packages
-from openwillis.features.video import (
-    facial_expressivity,
-    emotional_expressivity,
-)
-from openwillis.features.audio import (
-    vocal_acoustics,
-)
-from openwillis.features.speech import (
-    speech_transcription,
-    speech_characteristics,
-    speaker_separation,
-)
