--- openwillis/features/audio/__init__.py
@@ -1,5 +1,0 @@
-from openwillis.features.audio.acoustic import (
-    vocal_acoustics,
-)
-
-__all__ = ["vocal_acoustics"]
