--- openwillis/features/speech/__init__.py
@@ -1,13 +1,0 @@
-from openwillis.features.speech.speech_transcribe import (
-    speech_transcription,
-)
-
-from openwillis.features.speech.speech_attribute import (
-    speech_characteristics,
-)
-
-from openwillis.features.speech.speech_separation import (
-    speaker_separation,
-)
-
-__all__ = ["speech_transcription", "speech_characteristics", "speaker_separation"]
