+++ openwillis/measures/audio/__init__.py
@@ -1,0 +1,13 @@
+from openwillis.measures.audio.acoustic import (
+    vocal_acoustics,
+)
+
+from openwillis.measures.audio.speech_transcribe import (
+    speech_transcription,
+)
+
+from openwillis.measures.audio.speech_separation import (
+    speaker_separation,
+)
+
+__all__ = ["vocal_acoustics", "speech_transcription", "speaker_separation"]
