+++ openwillis/measures/audio/config/speech.json
@@ -1,0 +1,6 @@
+{
+  "result": "result",
+  "scale": "[panss]",
+  "text": "text",
+  "panss_string": "[Hi I am We are going to be spending the next 30 to 40 minutes talking about you and your reasons for being here. Maybe you can start out by telling me something about yourself and your background,Have you been feeling worried or nervous in the past week,Would you say that you are usually calm and relaxed,What is been making you feel nervous,Just how nervous have you been feeling,Have you been shaking at times or has your heart been racing,Do you get into a state of panic,Has your sleep eating or participation in activities been affected,Have things been going well for you,Has anything been bothering you lately,Can you tell me something about your thoughts on life and its purpose,Do you follow a particular philosophy any special rules teachings or religious doctrine,Some people tell me they believe in the Devil what do you think,Can you tell me more about this,Can you read other people minds, How does that work,Can others read your mind,How can they do that,Is there any reason that someone would want to read your mind,Who controls your thoughts]"
+}
