+++ openwillis/measures/audio/util/transcribe_util.py
@@ -64,7 +64,7 @@ def transcribe_audio(s3uri, region_name, job_name, language = 'en-US'):
 
         if status['TranscriptionJob']['TranscriptionJobStatus'] == 'COMPLETED':
             read_data = urllib.request.urlopen(status['TranscriptionJob']['Transcript']['TranscriptFileUri'])
-            
+
             response = json.loads(read_data.read().decode('utf-8'))
             transcript = response['results']['transcripts'][0]['transcript']
