+++ openwillis/measures/text/__init__.py
@@ -1,0 +1,5 @@
+from openwillis.measures.text.speech_attribute import (
+    speech_characteristics,
+)
+
+__all__ = ["speech_characteristics"]
