+++ openwillis/measures/text/config/text.json
@@ -1,6 +1,4 @@
 {
-  "result": "result",
-  "text": "text",
   "word": "word",
   "tag": "label",
   "tot_words": "tot_words",
